+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+## [1.4.0-beta.2](https://github.com/leaflockio/core-actions/compare/v1.4.0-beta.1...v1.4.0-beta.2) (2026-04-05)
+
+### Features
+
+* **deploy:** add use-netlify-build input to netlify action ([#78](https://github.com/leaflockio/core-actions/issues/78)) ([a204b4a](https://github.com/leaflockio/core-actions/commit/a204b4ac3ef895cb3131d052eb1dff74dc2b8882))
+* **hooks:** add eslint to node pre-push alongside coverage ([#77](https://github.com/leaflockio/core-actions/issues/77)) ([970c9e2](https://github.com/leaflockio/core-actions/commit/970c9e2301dcc1ea3c27419a9b03f65c1a987723))
+
+## [1.4.0-beta.1](https://github.com/leaflockio/core-actions/compare/v1.3.7-beta.1...v1.4.0-beta.1) (2026-04-05)
+
+### Features
+
+* **coverage:** add summaryFile field to COVERAGE_CONFIG_NODE ([#76](https://github.com/leaflockio/core-actions/issues/76)) ([0167bed](https://github.com/leaflockio/core-actions/commit/0167bedd2efd614151b410b3e36f36cdfd54389f))
+
+## [1.3.7-beta.1](https://github.com/leaflockio/core-actions/compare/v1.3.6...v1.3.7-beta.1) (2026-04-05)
+
+### Bug Fixes
+
+* **coverage:** per-metric node checks with JSON config and baseline ([#75](https://github.com/leaflockio/core-actions/issues/75)) ([db92303](https://github.com/leaflockio/core-actions/commit/db923037c5f68c39c75f7a9d7980e087d10d2ce0))
+
 ## [1.3.6](https://github.com/leaflockio/core-actions/compare/v1.3.5...v1.3.6) (2026-03-29)
 
 ### Bug Fixes